Add HP for divs recip rowexpanddiv colexpanddiv
