+++ lib/TileOps/tcolexpanddiv_template.py
@@ -5,11 +5,19 @@
 # THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND, EITHER EXPRESS OR IMPLIED,
 # INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT, MERCHANTABILITY, OR FITNESS FOR A PARTICULAR PURPOSE.
 # See LICENSE in the root of the software repository for the full text of the License.
-"""TileLang DSL template for pto.tcolexpanddiv"""
+
+"""TileLang DSL template for pto.tcolexpanddiv with IEEE 754 high-precision support
+
+Divide each column of src0 by a per-column scalar from src1[0, col].
+Semantics: dst[row, col] = src0[row, col] / src1[0, col]
+"""
 
 import sys
 from pathlib import Path
 import tilelang_dsl as pto
+
+# Import shared high-precision division algorithms
+from div_hp import _div_ieee754_f32_impl, _div_ieee754_f16_impl
 
 
 @pto.vkernel(
@@ -17,16 +25,32 @@ import tilelang_dsl as pto
     op="pto.tcolexpanddiv"
 )
 def template_tcolexpanddiv(src0: pto.Tile, src1: pto.Tile, dst: pto.Tile):
+    """Template for pto.tcolexpanddiv with optional high-precision mode."""
     dtype = dst.element_type
     valid_rows, valid_cols = dst.valid_shape
 
-    for row in range(0, valid_rows, 1):
-        remained = valid_cols
-        for col in range(0, valid_cols, pto.get_lanes(dtype)):
-            mask, remained = pto.make_mask(dtype, remained)
-            lhs = pto.vlds(src0[row, col:])
-            rhs = pto.vlds(src1[0, col:])
-            # TODO: 当前使用普通精度版本，后续需要添加高精度版本（vdivh）
-            result = pto.vdiv(lhs, rhs, mask)
-            pto.vsts(result, dst[row, col:], mask)
+    precision_mode = pto.get_op_attr("precision_mode", "DEFAULT")
+    if pto.constexpr(precision_mode == "HIGH_PRECISION"):
+        for row in range(0, valid_rows, 1):
+            remained = valid_cols
+            for col in range(0, valid_cols, pto.get_lanes(dtype)):
+                mask, remained = pto.make_mask(dtype, remained)
+                lhs = pto.vlds(src0[row, col:])
+                rhs = pto.vlds(src1[0, col:])
+                if pto.constexpr(dtype == pto.f32):
+                    result = _div_ieee754_f32_impl(lhs, rhs, mask)
+                elif pto.constexpr(dtype == pto.f16):
+                    result = _div_ieee754_f16_impl(lhs, rhs, mask)
+                else:
+                    result = pto.vdiv(lhs, rhs, mask)
+                pto.vsts(result, dst[row, col:], mask)
+    else:
+        for row in range(0, valid_rows, 1):
+            remained = valid_cols
+            for col in range(0, valid_cols, pto.get_lanes(dtype)):
+                mask, remained = pto.make_mask(dtype, remained)
+                lhs = pto.vlds(src0[row, col:])
+                rhs = pto.vlds(src1[0, col:])
+                result = pto.vdiv(lhs, rhs, mask)
+                pto.vsts(result, dst[row, col:], mask)
     return
