+++ test/tilelang_st/npu/a5/src/st/testcase/tdivs/cases.py
@@ -11,11 +11,26 @@
 """Single source of truth for tdivs ST test cases.
 
 vdiv only supports f16/f32 in TileLang DSL v1.
+
+Each case defines:
+  - name:        case identifier, used as subdirectory name and by main.cpp kCases[].
+  - dtype:       numpy dtype (e.g. np.float32).
+  - shape:       (rows, cols) — allocated tile dimensions.
+  - valid_shape: (valid_rows, valid_cols) — effective computation region.
+  - eps:         tolerance for numpy.allclose (atol and rtol).
+  - direction:   "src_scalar" (src / scalar) or "scalar_src" (scalar / src)
+  - precision_mode: optional, "DEFAULT" or "HIGH_PRECISION".
+  - test_pattern: optional, "normal", "precision_sensitive", "subnormal", "overflow"
+
+gen_data.py and compare.py both import this list to avoid redundant definitions.
 """
 
 import numpy as np
 
 CASES = [
+    # ============================================================
+    # Normal cases - basic functionality (DEFAULT precision mode)
+    # ============================================================
     # src / scalar direction
     {
         "name": "f32_32x64",
@@ -81,5 +96,150 @@ CASES = [
         "valid_shape": (256, 16),
         "eps": 1e-6,
         "direction": "scalar_src",
+    },
+
+    # ============================================================
+    # HIGH_PRECISION mode - src / scalar direction
+    # ============================================================
+    # Precision-sensitive ratios
+    {
+        "name": "f32_32x64_hp",
+        "dtype": np.float32,
+        "shape": (32, 64),
+        "valid_shape": (32, 64),
+        "eps": 1e-6,
+        "precision_mode": "HIGH_PRECISION",
+        "direction": "src_scalar",
+        "test_pattern": "precision_sensitive",
+        "ulp_tolerance": 1,
+    },
+    {
+        "name": "f16_63x64_hp",
+        "dtype": np.float16,
+        "shape": (63, 64),
+        "valid_shape": (63, 64),
+        "eps": 1e-3,
+        "precision_mode": "HIGH_PRECISION",
+        "direction": "src_scalar",
+        "test_pattern": "precision_sensitive",
+        "ulp_tolerance": 1,
+    },
+
+    # Subnormal numbers
+    {
+        "name": "f32_16x64_hp_subnormal",
+        "dtype": np.float32,
+        "shape": (16, 64),
+        "valid_shape": (16, 64),
+        "eps": 1e-6,
+        "precision_mode": "HIGH_PRECISION",
+        "direction": "src_scalar",
+        "test_pattern": "subnormal",
+        "ulp_tolerance": 2,
+    },
+    {
+        "name": "f16_16x64_hp_subnormal",
+        "dtype": np.float16,
+        "shape": (16, 64),
+        "valid_shape": (16, 64),
+        "eps": 1e-3,
+        "precision_mode": "HIGH_PRECISION",
+        "direction": "src_scalar",
+        "test_pattern": "subnormal",
+        "ulp_tolerance": 2,
+    },
+
+    # Overflow/Underflow boundaries
+    {
+        "name": "f32_16x64_hp_overflow",
+        "dtype": np.float32,
+        "shape": (16, 64),
+        "valid_shape": (16, 64),
+        "eps": 1e-6,
+        "precision_mode": "HIGH_PRECISION",
+        "direction": "src_scalar",
+        "test_pattern": "overflow",
+    },
+    {
+        "name": "f16_16x64_hp_overflow",
+        "dtype": np.float16,
+        "shape": (16, 64),
+        "valid_shape": (16, 64),
+        "eps": 1e-3,
+        "precision_mode": "HIGH_PRECISION",
+        "direction": "src_scalar",
+        "test_pattern": "overflow",
+    },
+
+    # ============================================================
+    # HIGH_PRECISION mode - scalar / src direction
+    # ============================================================
+    {
+        "name": "f32_32x64_hp_scalar_src",
+        "dtype": np.float32,
+        "shape": (32, 64),
+        "valid_shape": (32, 64),
+        "eps": 1e-6,
+        "precision_mode": "HIGH_PRECISION",
+        "direction": "scalar_src",
+        "test_pattern": "precision_sensitive",
+        "ulp_tolerance": 1,
+    },
+    {
+        "name": "f16_63x64_hp_scalar_src",
+        "dtype": np.float16,
+        "shape": (63, 64),
+        "valid_shape": (63, 64),
+        "eps": 1e-3,
+        "precision_mode": "HIGH_PRECISION",
+        "direction": "scalar_src",
+        "test_pattern": "precision_sensitive",
+        "ulp_tolerance": 1,
+    },
+
+    # Subnormal - scalar / src (scalar is normal, src contains subnormals)
+    {
+        "name": "f32_16x64_hp_subnormal_scalar_src",
+        "dtype": np.float32,
+        "shape": (16, 64),
+        "valid_shape": (16, 64),
+        "eps": 1e-6,
+        "precision_mode": "HIGH_PRECISION",
+        "direction": "scalar_src",
+        "test_pattern": "subnormal",
+        "ulp_tolerance": 2,
+    },
+    {
+        "name": "f16_16x64_hp_subnormal_scalar_src",
+        "dtype": np.float16,
+        "shape": (16, 64),
+        "valid_shape": (16, 64),
+        "eps": 1e-3,
+        "precision_mode": "HIGH_PRECISION",
+        "direction": "scalar_src",
+        "test_pattern": "subnormal",
+        "ulp_tolerance": 2,
+    },
+
+    # Overflow - scalar / src (division by small src values)
+    {
+        "name": "f32_16x64_hp_overflow_scalar_src",
+        "dtype": np.float32,
+        "shape": (16, 64),
+        "valid_shape": (16, 64),
+        "eps": 1e-6,
+        "precision_mode": "HIGH_PRECISION",
+        "direction": "scalar_src",
+        "test_pattern": "overflow",
+    },
+    {
+        "name": "f16_16x64_hp_overflow_scalar_src",
+        "dtype": np.float16,
+        "shape": (16, 64),
+        "valid_shape": (16, 64),
+        "eps": 1e-3,
+        "precision_mode": "HIGH_PRECISION",
+        "direction": "scalar_src",
+        "test_pattern": "overflow",
     },
 ]
